feat(aggregation): add fedmedian aggregation algorithm
Added fedmedian aggregation method. It computes the median of the weights,
enhancing robustness in the presence of Byzantine attacks by reducing the
influence of outliers. As some numpy operations change the weights type to
numpy arrays, a check was added in set_parameters() of lighting_learner to
convert them back to torch tensors if necessary.